+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - plugin import fix panel data
 - plugin import support geojson LineString
 - plugin categories new template panelPlaceEdit_cats_latest
+- plugin categories fix remove cat button
 
 ## v1.6.6
 - core support Instagram OAuth signup